Fix onConnet() is called twice for AugmentedAutofillService.
When AugmentedAutofillService name changes, onAugmentedServiceNameChanged()
tris to updateRemoteAugmentedAutofillService(). In getServiceForUserLocked(),
if we can not get the service from mServicesCache, the newServiceLocked()
triggers updateRemoteAugmentedAutofillService() which binds the service for
the 1st onConnet(). Then getServiceForUserLocked().updateRemoteAugmentedAutofillService()
will trigger 2nd onConnet(). This causes augmented cts tests flaky.

Due to getServiceForUserLocked() calls updateRemoteAugmentedAutofillService() if we
can not get the service from the cache, we can skip call update.

Bug: 144294363
Bug: 145136776
Test: atest android.autofillservice.cts.augmented
Change-Id: If6505cc05f032da4d20e86b976b7e5862f1b8e8f

Former-commit-id: 9cfd022783de9fe23c29fff50e2bfc79cb4a6a23