Updates the cached service while the package re-enabling
The autofill would flag to use the augmented service only when try
to start the autofill service which was disabled. The behavior keeps
the same even if the autofill service is re-enabled.
To fix the issue, updates the invalid cached service to reflash while
the default service package has changed.

Fixes: 146392487
Test: atest CtsAutoFillServiceTestCases
Change-Id: I44383a25ea4fb83cb38e4166272beb1ac2ebfa59

Former-commit-id: 9376b90d3f33a4b6ea1b6a8ad97f4f1d12d42a4e