Clean up the rollback when test is done
When multiple staged sessions are installed on a device which doesn't
support checkpoint, only the 1st will prevail. The test failed when
TestApp.A is not the 1st rollback to commit.

We need to clean up the rollback left by testRollbackWhitelistedApp so
TestApp.A is always the only and first one to commit.

Bug: 146991129
Test: atest StagedRollbackTest
Change-Id: I4afdace934b7b7e3b198169dc2eb0da66c65eb94

Former-commit-id: b13131f10c55a7b20fe3a558437445bb135ac94f