Merge ControlState into Control
All the information that was sent with a ControlState is now sent with a
new updated Control object. That way, overriding Type and changes to
name can be done inside subscribe.

Modified ControlsProviderService to encapsulate the callback service and
provide public methods to send information to SystemUI. In particular,
onLoad verifies that all Control passed do not have status information
and erases that information if it has any.

Add token for reusing the same Callback binder.

Test: build
Change-Id: I6ce5e9ad53899ea588b23e0f067b5d1f86a0bb18

Former-commit-id: ee57f49c7d8ab4d6c22608ae8de224b356f5b94e