Binder cache for Bluetooth getState().
Bug: 140788621

Test: A special build that puts the PropertyInvalidatedCache in verification
mode was loaded on the device.  Then one iteration of MPTS was executed.  No
cache inconsistencies were found and no SELinux violations (associated with
the binder cache) were found.  The number of cache misses was approximately
3% of the total binder calls.  All binder calls went through the cache.
Repeated the test with a work profile installed (to provide a second user
ID).

Change-Id: I2f1d0b6d61dedc5d1badb06a20c5932eca415904

Former-commit-id: 38265345e8c6d169001b910b2a2bd2724fe9241e