Fix /mnt/user/<userid> permission bits
Vold sets permission bits for /mnt/user/<userid> when mounting a
volume, zygote was setting different (more relaxed) bits for that
folder. Now we set the same (more restrictive) bits.

It is important to set it here as well because apps may start before
the volume is mounted and they should have the always have the same
permission bits for /mnt/user/<userid>

Bug: 135341433
Test: atest -c android.appsecurity.cts.ExternalStorageHostTest#testSecondaryUsersInaccessible

Change-Id: I87af1b5e63601b38443715fced8b20693b9cf5e6

Former-commit-id: 1fc3355e8111986b5acc79ac41515ab668f3b946