Preserve install initiatior package on uninstall.
Previously when an installer package was uninstalled we removed all
reference to it in InstallSource.

Now we retain the initiator package name (but not installer or
originator). We set a flag to remember it has been uninstalled, and we
also store its signatures which would otherwise cease to be
available. Update persistence logic.

Also minor refactoring of InstallSource#create to separate the normal
case from those that need to set every field.

This information is not yet exposed in the API, but it will be.

Test: atest PackageManagerTests PackageVerifierTest
Test: Manually verify XML contents before & after uninstalling an installer.
Bug: 134746019
Bug: 146555198
Change-Id: I1452402de554b916e5f79bd386c5857d4860ba12

Former-commit-id: 72b7e673c79c5d84a4871a6f6e28d6b79071fc74