Add a copy of Rlog for the telephony mainline module
We can't expose android.telephony.Rlog directly to the telephony
module as @SystemApi because we can't protect it from app code.

So instead, expose a lower level API from the Log class to the
telephony module with client=SystemApi.Client.MODULE_LIBRARIES so
it can only be accessed by boot classpath code, and then create
a copy of Rlog in the telephony module, which internally use the
aforementioned API.

Note:
- Support for MODULE_LIBRARIES is not fully in yet, so for now
the Log.logToRadioBuffer() API is still hidden.
- Once all the infra is in, we can expose logToRadioBuffer.
- Once we mark it with MODULE_LIBRARIES, ART will protect it
against app code, so no further protection will be needed.

Bug: 144374158
Test: boot
Change-Id: I5be0497a17d00152286fe3351bf58e75b675ceb6

Former-commit-id: 57e56f2d3cfe9eb9e73198e3ed2b450f9df9d933