Move a bunch of row setters into an init method.
We have a bunch of setters for stuff that is only initialized once on a
row. So let's just go ahead and move these into an initialize method so
it's more readable and harder to mis-use.

There are a few others that are used in tests that should probably be
moved into this method or just moved out of ExpandableNotificationRow
entirely.

Test: builds
Test: atest SystemUITests
Change-Id: Ifa55ca19fa2fd1f2030caa0676e6aa3e1713f134

Former-commit-id: 5182c9f5d734205b1d266306d8bdc1a24953bc8c