Move a bunch of row setters into an init method.
We have a bunch of setters for stuff that is only initialized once on a
row. So let's just go ahead and move these into an initialize method so
it's more readable and harder to mis-use.

There are a few others that are used in tests that should probably be
moved into this method or just moved out of ExpandableNotificationRow
entirely.

Test: builds
Test: atest SystemUITests
Change-Id: I730096a6c8313a82226116a6c88c729bf450f7b5

Former-commit-id: 8271c2726917e63885519be141d84557b4632f44