Allow test builds to optionally skip verification
Currently, every build a developer creates will be verified
against the Play Store verification system. For developers, this
is a completely useless step, takes up resources, and eats into
iteration time.

This CL disables verification for debug development builds that
are reinstalled over ADB. This keeps users who install OTA safe,
and lets devs avoid the cost of verification.

There is no reduction in safety for the end user as:
1) Verification can only be skipped when installing over ADB and
   the "-t" as well as the new "--disable-verification" flags are
   specified ("-t" isn't strictly necessary).
2) The user can already disable verification over ADB by simply
   setting a flag.

Test: atest PackageVerifierTest#testAdbSkipVerification
Bug: 138672462
Change-Id: Id0bef5126f2fb49a8e30fc235193636a1c2bab2e

Former-commit-id: 750b5ed85cfd2c8598634ca96a75aba115130e13