Remove resource overlayable configuration
Resource overlay configurations are an "All or Nothing" situation
where we either have to list all overridable settings or none.

We're currently not in a place to find all possible overlayable
settings, and overlayable configuration isn't a requirement for
R, so I'm removing the single defintion I added in an earlier
CL to restore the previous behaviour of allowing all settings to
be overriden by vendors.

Fixes: 147408832
Test: Rebuild for Flame, manually confirm overridden setting is correct
Change-Id: Idccdee8197529b09fe2cb3e8701550c3c896f46e

Former-commit-id: 114cb5bdf7d938c764f8ead2f57bbc684d2c7c62