Throw exception out of grant/upgrade permissions callback.
PermissionControllerManager now uses AndroidFuture for callback, and
it swallows exceptions thrown in the callback, so it is not going to
kill the system server and main thread waiting on CountDownLatch will
never be unblocked. This change uses
AndroidFuture.completeExceptionally() to correctly pass out the
exeption and kill system server in that case.

Also added logging user ID to figure out which user is causing the
exception.

Bug: 147418116
Test: presubmit
Change-Id: Ib5a9073b0ef4d43232f9e4266ac072ad00f6afe7

Former-commit-id: 72ddbffcd5bf25070fd7cc020b013ab011f1cfd0