Delete the NetworkFactory running-on-correct-thread assertion.
The change to make NetworkFactory use NetworkProvider included an
assertion that the onNetworkRequested and onRequestWithdrawn
methods were called on the factory's Looper.

This is not really necessary since it cannot happen in the
current code unless there is a bug in Looper. It also makes it
impossible to test this code using TestLooper, which dispatches
messages on the test thread instead of on the looper thread.

So, delete the check.

Bug: 138306002
Test: atest FrameworksWifiTests now passes
Change-Id: I37449efb50287c683312f0b19909bd4c61b38585

Former-commit-id: 9ded26e2412a45a82d37fbd31fee5ce5155321be