Include an AppOp to allow dialer role app to record and play audio into calls
This change includes an AppOp that will be used to grant access to telephony call audio (RX and TX) to the default dialer application (assigned to the Dialer role).
A new permission was created so only apps that also require this permission will be granted. It will also be used to keep track of usage (audit).

Test: Compilation and manual verification
Bug: 135197853

Change-Id: I5d042f667ca5fe64f8048b8aff7b98f51b3c1500

Former-commit-id: e96b5668b35b14b98e18a4d64d7aadb7fa071f7f