Do not migrate use full volume shared preference
Use full volume has been renamed to override DnD and
the toggle is now only available for certain countries. We
intended not to provide the ability to migrate this shared
preference as it might cause issue for users who already
changed the preference. After mainline migration, override
DnD should be always reset to default value of that country.

Test: Manual
Bug: 147426188
Change-Id: I5746fe32674131b474e784bd5dd62ecff76386d5

Former-commit-id: bcbe46816c7febb2641413e363cb634f7fb8f18b