Introduce fixed rotation transform
It allows activity to be shown in an orientation different
from the current display. This initial change only handles
the case when transiting to another activity with different
orientation.

The opening activity and its process will be updated with
rotated configuration. The rotated display frames and insets
are also associated to the activity, then the activity can
be layouted in the rotated bounds. The surface of activity
is rotated to fit current display. And the display will be
rotated seamlessly until the transition is done.

The function is disabled by default. It can be enabled by:
adb shell settings put global fixed_rotation_transform 1

Bug: 143053092
Bug: 146415879
Bug: 147469351
Test: atest DisplayContentTests#testApplyTopFixedRotationTransform

Change-Id: Iea15c6fc68133728e5fcda87ba4efd1850fdf5d4

Former-commit-id: 6f548e99cb847168738c10cc60c38aea0a8077da