Store the correct rollback id for logging on boot complete
Previously, RPHO expected there can be only one train staged for
installation. Hence, it expected there can be only one rollback
committed at any time. As such, it only saved the latest comitted
rollbackId and on boot complete, used that rollbackId to extract
ModuleMetadata package and log.

Now that we allow multiple staged installation and on native crash
during boot we commit more than one rollback, we cannot blindly store
the latest rollbackId for logging. The logging is only meaningful for
mainline trains, that is, when when we are rolling back a train that
contains ModuleMetaData.

In this CL, instead of storing the latest comitted rollback id, we store
the rollbackId of train that contains a valid logPackage, i.e,
ModuleMetaData package.

Bug: 141843321
Test: atest StagedRollbackTest
Change-Id: I94a3330c8abbc9470f4a698c3333c55a044cbd10

Former-commit-id: d2f50759766dca6c58f2472d36072e8484ce05ea