Snapshot and restore CE apex data directories.
This change extends the existing mechanisms for snapshotting and
restoring CE data for APKs to also support CE apex data.

StagingManager now calls through to RollbackManager to perform the
snapshot/restore (at the same time as it does for APK-in-APEX,
added recently).

For APEXes, the actual snapshot or restore is performed by apexd instead
of installd. Tracking of pending backups and restores is performed by
RollbackManager in the same way as for APKs.

Bug: 141148175
Test: Manual install/rollback of an APEX with dummy files in CE data directory.
Check that snapshot and restore occur as expected.
Test: atest CtsStagedInstallHostTestCases
Change-Id: I82e1818f927881ea774148d97ca1981a447b598d

Former-commit-id: 7bbd8b471f63514c82e51d4261f1bbc40389d745