Add build target for permissions text file
This CL adds a genrule that generates a text file that contains all of
the normal / non-signature permissions formatted with one permission per
line.

This can be used by tools that require a list of permissions known as
part of their build. Without this support, it would be difficult to
write the build rules for such tools as it is not possible to access the
frameworks/base/core/res/AndroidManifest.xml from Android.bp files in
other directories.

Test: Built locally and manually verified the file
Change-Id: I01046e5faa1f0a81c3d2ee82643094d4cf82f221

Former-commit-id: 907b70ae3d82133cd2b78e22edb89c096ca3c21f