Clean framework-aidl-export-defaults up
The defaults module specifies AIDL paths that are exported to the
clients of framework or framework-minus-apex libraries. Therefore, it
doesn't need to be inherited to framework-annotation-proc whose purpose
is just to run the annotation processor and isn't linked to anyone.

Also moved the 'required' property to the other defaults
'framework-defaults' since the property has nothing to do with exporting
aidl directories.

Bug: N/A
Test: m
Change-Id: Ied8522eec5c03d3162e6ae22a08e21f31dc528d2

Former-commit-id: 392729fcc987e1bc341c420559e745694b897ed1