Add an unsuspend button option to SuspendDialogInfo
The suspending app can specify if they want to unsuspend the app on
neutral button tap. This needs to be done by the system so as to
maintain continuity of the users activity tasks.
The system then also needs to communicate to the suspending app when
such an unsuspend has taken place so it can keep any of its internal
data consistent with the current state.

Test:  atest FrameworksServicesTests:PackageManagerSettingsTests
atest FrameworksServicesTests:SuspendDialogInfoTest
atest GtsSuspendAppsTestCases:SuspendPackagesTest

Bug: 134962563
Change-Id: I587b2643eb41519b82b8e2b371a466a9fb650d4b

Former-commit-id: 0cb5713fd7faf31aa41d8a1c8300268c10833125