Allow an app to see its installing package.
An app is automatically granted visibility to its installing app and
its install initiating app (provided it has not been uninstalled), but
not the originating app.

Add corresponding unit tests.

Make sure we update the apps filter if the installing app
changes. (The others can't without an install or update.)

Fixes: 147795820
Test: atest PackageManagerTests AppsFilterTest
Change-Id: I5e5e03636b97c1d56dbcc6752fe2a2bb70d28925

Former-commit-id: a547c5c63cdb6cf8149875c7c3fad18d69bc976f