Remove annotations from MmTelFeature.MmTelCapabilities.toString
Object.toString() is a public API. Marking a method overriding it as
@SystemApi or @TestApi doesn't make the method a system API or a test
API. In fact, it breaks CtsSystemApiAnnotationTestCases which ensures
that a Java symbol with @SystemApi annotation is in the API signature
text.

Removing the unnecessary annotation to fix the test. Note that the API
signature file is not changed with this.

Bug: 147812912
Test: atest CtsSystemApiAnnotationTestCases
Change-Id: I47c07f072f5cffc64a8867f27889745065f73fd1

Former-commit-id: 04701fe9f4273df650c2cb8837acca1c75ea82ef