Mark storage encryption functions as deprecated.
The setStorageEncryption/getStorageEncryption methods do nothing and
never have they had any impact on the system.

All setStorageEncryption does is store a value that can be retrieved by
getStorageEncryption. Otherwise, it calls setEncryptionRequested
internally which has an empty body.

Mark them as deprecated so they could be removed in the future.

Test: that it compiles
Bug: 144412716
Change-Id: I007a1acf80ddb7cba9717f2e541f029c5f379d11

Former-commit-id: d4e25d49b0a2070532236e4e2d0cce0c72430d22