Expose RESULT_RECEIVER_CONTROLLER_KEY as @SystemApi.
It's used by a @SystemApi as key of it's return bundule. So we
should expose it. Also it's currently linked to BatteryStats.
RESULT_RECEIVER_CONTROLLER_KEY which makes no sense as they are
functionally independent and BatteryStats.RESULT_RECEIVER_CONTROLLER_KEY
is @hide. So also make it a hardcoded String.

Bug: 148116922
Test: build
Change-Id: I7099048c126e88f75cf5bd7e779ddfe923cc1c02
Merged-In: I7099048c126e88f75cf5bd7e779ddfe923cc1c02

Former-commit-id: c27a3f37b1477afb73ce0412434547cfcf7bb334