Fix javadoc @link tags on @ChangeId constant.
In future, we will generate DAC docs form these javadoc comments. But
due to the way that this works, the docs generation loses context of
constants in classes that are not part of the SDK, meaning we can't
refer to imported classes directly. To keep the links working, class
references must be fully qualified.

Bug: 144927670
Test: m
Change-Id: I059698f224ce060600f69ef6c816d5a81d143968

Former-commit-id: dbc20a773f79b1ee9719e94354745dc2535b5d1b