Switch to compositor shadows for PiP and Freeform windows (disabled by default)
Shadows are generated on PiP and Freeform windows by setting an
elevation on the window's DecorView. The elevation is passed back
to wm as surface insets to generate a large enough surface for the
client to draw its content and render shadows around it.

This cl introduces a new settings which overrides the setElevation
call in DecorView and sets a shadow on the task surface. The
elevation of the DecorView and the surface insets will remain unset.

Add a surface crop to the task surface so we have proper bounds
to draw shadows.

Test: go/wm-smoke
Bug: 136561771

Change-Id: Icc45e4fac3d495df5272d2e5f37163b9e7caeb4c

Former-commit-id: f6ef1c7071f3ca7250cc6d9078155b9d6b1d9fa8