Epsilon comparison in RefreshRateRange constructor.
In DisplayModeDirector there may be votes for refresh rate ranges
which are within EPS of each other but do not intersect. For example
[0, 60] and [60+eps, 60+eps]. This CL adds an epsilon check in the
constructor of RefreshRateRange and a unit test for the described
scenario.

Fix: 148310029
Test: atest DisplayModeDirectorTest
Test: Manual with a device with PEAK_REFRESH_RATE = 60, from an app
setting the  preferred mode to a 60fps mode (which appears as 60+eps
on the device).

Change-Id: I24e225edc0d7e388897e9e30f5bce858ac56cb06

Former-commit-id: fb3879f221f17ce100bc606d2c89868030169a2f