Show all licenses when multiple.
Unreadable partitions like recovery have their notices copied to
/system/etc/NOTICE.xml.gz or to /vendor/etc/NOTICE.xml.gz or to both.

If the notice requirements for a library change between building a
vendor blob and building a sytem image using that blob, the UI hides
one of the licenses. This makes the build system harder to understand
for developers and could cause other hassles.

Test: Manually using 2 devices.
Test: atest

Change-Id: I018d2a1949d11b201e4eeb23f186be8d244e161d

Former-commit-id: 07abc181a878754032253a0befcd1e7a289114b6