Add FrameworkStatsLog
Add a genrule for FrameworkStatsLog and it to frameworks.jar

As part of statsd becoming a Mainline module in R, autogenerated
StatsLog.write() calls are going away and replaced by *StatsLog.java
that is autogenerated for each module.
This CL adds autogenerated FrameworkStatsLog.

Follow up CLs will replace usages of StatsLog.write calls with
FrameworkStatsLog.write.

Bug: 145952197
Test: m
Change-Id: I6ba6a67e019a3b5cffd936dad7ce9db6c8c51632

Former-commit-id: 04ec2529f7409c87a96d180e12f9fff9bffc97e2