Specify explicit actions for bubble / unbubble
It's not always safe to flip the channel setting because the notif might
be allowed to bubble, but is not currently displayed as bubble. In that
case the button would say "show as bubble" but the code would flip the
setting to false, doing the opposite of the user action.

This CL specifies actions for bubble / unbubble and only commits the
channel change if the setting actually changed.

Test: atest NotificationConversationInfo
Fixes: 148425750
Bug: 148424396
Change-Id: I6f561cf388d2595df53ac41b9a183f6c50976d21

Former-commit-id: d7fcf05d3f8ee7193e3b33724085d0e72e9cfa18