Update mReadOnlyNotifList in ShadeListBuilder
So that mReadOnlyNotifList always tracks mNotifList, even after
switching mNotifList and mNewNotifList

This CL also updates ShadeListBuilderTest to catch bugs like this.
Mockito keeps shallow references when verifying mocks, so we can't rely
on verifying the passed list after the test is run (since the list is
recycled within ShadeListBuilder). Instead, make a deep copy of the list
we receive in the listener and verify this list for equality.

Test: atest ShadeListBuilderTest
Fixes: 148421960
Change-Id: I85eb439f77686c67b365f04585da8fc54cb854ec

Former-commit-id: 975329730508ffea302c5dc06603c8b42de29b25