Kill app when OP_REQUEST_INSTALL_PACKAGES is changed.
When an app requests the special access permission to be able to install
packages, and it is granted, we need to kill the app, in order to
provide the correct (more permissive) filesystem views. Likewise, when
the app op is taken away, we need to kill the app, to prevent it from
making modifications it is no longer allowed to make.

Bug: 134706060
Bug: 133504844
Test: Use Epic installer; verify it gets killed when the app-op is
granted.

Change-Id: I73c5ce56ee24e5c864d7c9280b83fc3854701a97

Former-commit-id: 28ec85a35e67494ac6d34773815e8d87abce9b62