Use FrameworkStatsLog instead of StatsLog
As part of statsd becoming a Mainline module in R, autogenerated
StatsLog.write() calls are going away and replaced by *StatsLog.java
that is autogenerated for each module.
This CL replaces some usages of StatsLog with
FrameworkStatsLog.

Using FrameworkStatsLog here for now but this will have to change to a
specific *StatsLog for Jobscheduler once Jobscheduler becomes a Mainline
module.

Bug: 145952197
Test: m
Change-Id: I1090ade2abf3f64e0ae4bf662794e46ea7229c96

Former-commit-id: e9aec32f977ab3f2bd2cccecd1124a059c912ebd