Dark theme helper
Since adding custom and auto night mode, applications have been
making mistakes in checking if the device is in dark theme.

The API for checking dark theme is not very clear. They use
UiModeManager.getNightMode and checking if it is a yes or no thinking
that it refers to Dark theme. Where it is the service scheduler modes.

The most famous example of such mishap is the gboard app (b/145820343) and the dialer
(b/148142709).

We had to test all google apps that implement dark theme on android.

Test: atest UiModeManagerTest and Config tests
Fixes: 148606412
Change-Id: I45dbd9daa057b14d5dd2e88115f666db99d19145

Former-commit-id: 5948319efc57115eb479ebdbdb7cefc32417bd9c