Fix experimental bubbles from HUN'ing when they shouldn't
I changed this to onNotificationAdded so that we could extract
the avatar generated for the messaging template for experimental
bubbles, however, that breaks how notification interruption state
provider (NISP) works (we get onNotificationAdded after it goes
through NISP that to check if the notif should HUN). It's technically
more correct to use onPendingEntryAdded here, so lets do that.

If the person object doesn't have an avatar we'll end up using
the small or large icon from the notification instead, this at
least works for most apps I've tried with experimental bubbles.

Test: have no hangouts bubbles, get a hangouts bubble, note that
      it doesn't HUN
Fixes: 148609860

Change-Id: I6e19e49b8b5f13969c30410450826bf480fb9d34

Former-commit-id: f9439ab4072c71b294167524fdddb5cbe39a5923