Add init/cleanup hooks to notif listeners.
As we refactor more state out of NotificationEntry, we need a place to
initialize/delete this state in its respective manager. Add/remove is
not appropriate for this as some add callbacks will call into other
managers which may have not had a chance to initialize their state yet
since we have no guarantee of ordering.

So instead, we break out a init and cleanup callback from add and
remove respectively where listeners can use it to do any initialization.

Test: atest NotifCollectionTest
Change-Id: If2233f8f4058c91dcf3d6ede61deae656e47c728

Former-commit-id: 59609424f72ffb34f7ae482c0da43f0e1139ea6e