TaskOrganizer: Don't set any position at all.
For the first implementation of SysUI driven PIP we won't do
any reparenting, and so there's no leash on their side. This means
we don't wan't to set a 0,0 relative position. For the Tiles case
we also don't want to do that. In Bubbles we do want this but we can
just set a 0,0 position when reparenting to the leash. The common
guarantee we need is that the WM wont mess with it!

Bug: 146594635
Bug: 139371701
Test: Existing tests pass.
Change-Id: I08b8433c88fd8911ff1aaa96a123369e914587da

Former-commit-id: f7e751a247fe1913af456380f976e06549dee72e