Allow the FUSE daemon access to Android/ on non-sdcardfs devices
Today, the FUSE daemon has access to Android/ because it has a
writable sdcardfs view.

On devices without sdcardfs, it will not have this view and will need
the two new ext_data_rw and ext_obb_rw. In future cl, we will give
these gids to access_mtp, so we optimistically add the FUSE daemon in
the condition for access to Android/

Test: m
Bug: 148555237
Change-Id: Ia4dd36236d302842c89f5cd13b3235805f235224

Former-commit-id: 6462805c595aa88ec95dbd4c754371e5edc7ce16