Don't build framework-annotation-proc with framework.jar
Previously, framework-annotation-proc was built with framework.jar in
its classpath.  This was because sdk_version is not set for the module,
in which case the build system automatically add framework.jar to the
classpath.

This is wrong because the classes from framework.jar are also provided
via the srcs property of framework-annotation-proc. With this setup, a
build error could happen when there is a jarjar'ed class. The class is
jarjar'ed in framework.jar, while the same class in source files isn't
jarjar'ed yet when compiling it.

To fix the problem, remove framework.jar out of classpath by setting
sdk_version to core_platform and provide additional libraries via libs
property.

Bug: N/A
Test: m
Change-Id: I00974c76aaa0ce40384ffa6bf13d4f0ba5e37954

Former-commit-id: 963ee178c993b77fe75ac4d104b4fa16c7bcdc08