Correct a permission check / add a test
This change corrects a permission check in TimeZoneDetectorService
which is now covered by a test.

A TestHandler common to the timezonedetector and timedetector classes
has been extracted and made less flaky.

There are some changes included to make the equivalent time zone and
time classes consistent and to improve method names.

Bug: 140712361
Test: atest com.android.server.timezonedetector
Test: atest com.android.server.timedetector

Change-Id: Ic380ede6c7276d6b80f0fc74b30bba8c89c07fcc

Former-commit-id: cce4dcd300efe1268e74c311807c719aa678b3f0