Fix ClassCastException from toArray().
toArray() returns Object[], which can't be cast to Foo[]. Switch to
toArray(Foo[]) instead.

Fixed one instance which caused test breakage, and then all the others
I could find.

Also make sure PackageInstallerSessionTest runs in presubmit. It's a
cheap unit test, and tests are more useful when they run. (This is
what I was trying to do when I discovered the test was currently
failing.)

Fixes: 148858333
Bug: 137951074
Test: atest PackageInstallerSessionTest
Change-Id: Ie0142d46f1a34cf2072aaf81b3a5f8aa76d0c1f2

Former-commit-id: ee59d5462c9693b2375ecb1dcf77c2c55cd7b122