Set defaults for cached and empty processes.
Values for cached and empty processes are read from DeviceConfig.
There's no reason to update the values when Settings constants change,
so removing that call.

Also, since we switched to loading DeviceConfig flags in a loop, the
defaults would never be set for cached and empty processes, so fixing
that by explicitly setting defaults, which is consistent with all the
other constants in this file.

Bug: N/A
Test: Android builds
Change-Id: Iec1973d9c7eb5440311a0a51d5616c87ee677f3f

Former-commit-id: 444ad6a7773412b33719e125f96af4ba1b2301b0