Add null check for voiceInteractorComponentName.
Voice Assistant can be disabled on a user-level. Adding this null check
prevents crash for when Assistant is enabled for the system in general but
disabled for a specific user.

Bug: 149112015
Test: Manual -- Change config_disableLockscreenByDefault to true to
emulate the environment in which Volvo observed the bug. Verify that
switching to user0 through adb shell switch-user 0 causes the same crash.
Verify that the crash does not happen with the new null check.

Change-Id: I5b8ede1e5bd8c1bc047bc6d6220b425dea8f50ea
Merged-in: I5b8ede1e5bd8c1bc047bc6d6220b425dea8f50ea

Former-commit-id: d445aaf1e2e4c9e98f03d0726cd513a34952a693