Changes to favorite management screen
This changes are to include the new UX flows:

* The list of controls is split by zones. All controls that do not have
a zone (blank or null) are shown at the beginning of the list. Zones are
sorted "alphabetically".
* Controls are split into Favorites and All, with Favorite controls
appearing twice. Toggling the checkbox in a control from any of the two
lists will change the state and make it appear/dissapear from the
favorites list.
* Changes to favorites are only saved on Done.
* Favorites can be rearranged by dragging and dropping.

Test: manual
Test: atest
Bug: 149138395

Change-Id: Ib5fdbc66cafe03e95855eb6c16c27127afa8d565

Former-commit-id: 9aa23af654b262532c0613180509ef3e563dfeb9