Use associated display to get real size
The display may be removed when initializing navigation bar
for the display (add and remove quickly). Display manager
will return null by the removed display id, then NPE.

By using the display context, it makes sure the display is
non-null because the context is created from createDisplayContext.
The getRealSize just does nothing if the display was removed.
And since the window will be removed soon, the states of these
pending remove components are not important.

Bug: 148952431
Test: atest MultiDisplaySystemDecorationTests
Change-Id: Iea5a0cab6e033418a93900627622034cd491b0fa

Former-commit-id: 410d9751035b2a268bc7b2f895d384bbb62bbc28