Don't auto-dismiss children with certain flags
Typically, when a group summary is dismissed, NotificationManager will
also dismiss that group's children. In SysUI, we anticipate this change
by optimistically dismissing those children on our side before we get
confirmation from NoMan.

However, there are exceptions to this auto-dismissal behavior, namely if
the child is part of a foreground service or a bubble. This CL adds in
those exceptions to the SysUI side.

Change-Id: Iff03846501af3736897bb61fe825cca911d7584a
Test: atest

Former-commit-id: 2e6396664dc7258dc3de9b1f38f7dd71d15eb007