Use #getRollbackForSessionLocked to search for the rollback (3/n)
Each session is allocated with a unique id. No matter it is a parent
session or a child one, a staged session or a non-staged one. For a
given session id, we can correctly locate the matching rollback without
checking #isNewRollback.

In the end, we will be able to remove #getNewRollbackForPackageSessionLocked
when the refactoring is done.

Bug: 149069841
Test: atest RollbackTest
Change-Id: Id7166def31d53ab33dc5212b038a135076c6cb8c

Former-commit-id: 06650e7326efe5c67d989f3d2bc49cbd86c79324