Make sure the control target has a control
The CL only allows a target to be the control target if the insets
source provider has a control. This prevents the arrays in
InsetsStateController.mControlTargetTypeMap contains a type which
doesn't have a control. If an insets source provider haven't had
a window yet while we are updating control for target, we let the
target be the pending control target, and then use it to update again
while the provider get the window.

Bug: 149121559
Test: Manual test
Change-Id: I3c63e892f870f4daca667106d935e05423dfece8

Former-commit-id: 3b3537e0a18844b479f9585eb9589c61d9418665