Clean-up Bounds Animator code path (92/n)
Use configBounds to set the configuration bounds exposed to the app and
displayedbounds to set what the user see happening with the surface.
This is clearer than the concept of stack bounds vs. task bounds. which
we won't be able to use for tasks/stacks that are only one level deep.

Bug: 10163964
Test: Pip works and existing tests pass.
Change-Id: I986a5098cd81ea1c455c44eb2f704d803cf10b50

Former-commit-id: ea5e87f66742a6f8a8c80d94cc4879c9429aba4c