Don't override activity display adjustments with app config
When process configuration was applied on the client side it
accidentally applied an override to display adjustments in resources
for all ResourceImpl objects. This resulted in resources of
activities having incorrect display adjustments and reporting
incorrect display size.

This CL fixes the issue by updating the applied config with the
activity override. It also removes
ResourcesManager#applyNonDefaultDisplayMetricsToConfiguration, which
is no longer needed on the client side since the server-side is fully
multi-display aware.

Bug: 148639826
Test: ActivityThreadTest#testHandleConfigurationChangedDoesntOverrideActivityConfig
Test: AppConfigurationTests#testDisplaySizeInSplitScreen
Change-Id: Ib3ee007bcccd11451556b66274de4257cba082d4

Former-commit-id: 8efbbe751fa5167d2a776c3acd24d9eb644ad4a9