[Passpoint] Fix connected network not showing in Wi-Fi picker
Newly added network by OSU does not appear in Wi-Fi picker even
when connected.
Root cause: newly added unique identifier included the certs and
key material in the hash, but after the profile is installed,
the certs and keys are moved to keystore and removed from the
configuration, hence the change in hash.

Bug: 149322822
Test: Manual OSU connection and provisioning successfully,
verifying connected network appears in Wi-Fi picker
Test: atest PasspointConfigurationTest

Change-Id: Idec09bc77cb3562bb6bfe9e586a3aa7b2dce85bd

Former-commit-id: 697e02a3b80c8f3eb897ba5918e3071ef55262a5