Don't return early on error (1/n)
Continue the call flow even when #enableRollbackForPackageSession fails
so #completeEnableRollback has a chance to delete a rollback when any of
the child sessions fails to enable the rollback for the package.

Without this CL, the rollback data won't be deleted until next reboot.
Since storage cost might be heavy for the rollback data, we want to
delete it as soon as possible without waiting for reboot which might
happen much later.

Bug: 149352598
Test: test RollbackTest StagedRollbackTest
Change-Id: I4aeda1f2447e1b12353d2e3866408aab67a637eb

Former-commit-id: b3ea9a61d4e847496c90708ab53655fed97ccfb7