Store hard refences in a static context and pass through only weak references during the Content Capture Sharing.
Motivation: if the remote app is killed, we don't want a possibility of
system server holding a stroing reference (through a reference chain)
to large objects in that app. Therefore what's send in the binder has to
be a weak reference. And we will store a hard reference to those objects
in the client app's static context.

Storing hard references to objects in system_servier is less critical, because that is not going to be killed.

Bug: 148265162
Test: covered by CTS tests
Change-Id: Ie561aab6019d191cf8659fb350e045089e7781ed
(cherry picked from commit 4daacefe8b2b6291ea32117e8ba0b38156677abb [formerly 13f65b29745d306351c666c4f9f4a2d701513fe5])

Former-commit-id: 5532c77904352b80c44d0a783532b418f497f74e