Use event text for partial accessibility state change in quicksettings
For the accessibility state description change event, if only part of
the state description is changed, we can put the changed part into event
text so that accessibility service knows the changed part.

Test: tested with the wifi tile and blooth tile. When state changes in
the on state, only the changed part is announced in talkbacki when the
focus is on the tile. The "on" state is not announced repeatedly when the
tile remains in the on state. All other behaviors are the same.

Change-Id: I05ebe806e7887ba3d0496b3819fac2a7184813cc

Former-commit-id: 1f45b58fdb3225ff679a9a8f0d349a4cf49481bc