Broadcast screenshot intents in the foreground
By default, intents are broadcast on a background queue, which can
accrue a lot of backed-up tasks especially when the phone was
recently rebooted. In the case of the screenshot action shade, this
causes a long pause (up to ~minutes!) between tapping on an action
chip and seeing the app actually open.

This change adds a flag to the intents to cause them to be queued
in the foreground. In the notifications flow, this flag gets added
in NotificationManagerService (so this is preserving behavior). In
addition, these intents are user-generated (by tapping on the chip
or screenshot preview), so it's reasonable to send them in the
foreground.

Bug: 149481318
Test: manual
Change-Id: Idfcb8ade3fee232b140299585296e85093eeb1fe

Former-commit-id: ffb189773c0cdae08c8f5d88bac8a70f1b9a2136