Add "enableGwpAsan" tag to android manifest.
This tag can be used to enable and disable gwp-asan on applications and
their components. The default setting is disabled, except for the system
applications which are enabled with a small probability (approx. 1% of
processes).

Enabling gwp-asan can be used to crowd test an app.

This change includes a compat feature that overrides the default value
of enableGwpAsan for an app.

Bug: 149991821
Test: atest CtsGwpAsanTestCases

Change-Id: I5eb647c517e5487b1d1eed6b9a43242490f19289

Former-commit-id: db076c83ada31c8944ee690d14d1c63ffc5e23b4