OMS: block until settings are persisted
Do not schedule a job to write the OMS settings file on a background
thread any more. Instead, wait for the file to be written before
returning. This is a step towards a transactional OMS API, where it is
vital that the file is always up-to-date in case a transaction fails and
the in-memory settings need to be rolled back.

Also: move the PackageReceiver off the main thread to prevent blocking
the main thread. (AIDL calls are already processed on one of the binder
pool threads; UserReceiver is kept on the main thread.)

Bug: 119916381
Test: atest FrameworksServicesTests:com.android.server.om OverlayDeviceTests OverlayHostTests CtsAppSecurityHostTestCases:OverlayHostTest
Change-Id: I2452f789af1353fdde729bdc841b3417a2048e99

Former-commit-id: 038b24eb67d0416869bd0130af9bdce8d0ba3f6e