Compute Wallpaper scaling on the server
Image wallpaper uses DisplayInfo to compute the wallpaper scaling.
When the wallpaper has a fixed rotation transform, its DisplayInfo do
not correspond to its transformed bounds.

We can't use the Configuration's bounds either because the wallpaper
token is created on the server side and we do not have a way to
propagate the changed configuration to a token created from the server.

Bug: 148000485
Test: atest WallpaperControllerTests#testWallpaperSizeWithFixedTransform

Change-Id: Ie80592ae868c980faddece17b94cd34178b1de0e

Former-commit-id: f79baeaa200d0a6a6d0774b1e9261d2ae97e2d99