Allow TestableLooper to be asserted as main looper
Tests are run from the TestableLooper, but we still want to pass
Assert.isMainThread calls. Therefore, allow SysUI tests to assert the Testable
Looper is the main looper.

We cannot set sMainLooper to TestableLooper because there may still be
messages in the MainLooper's queue that will get evaluated in the
future.

Test: atest SystemUITests
Fixes: 149740127
Change-Id: I0c76763766f5ed4a85d99aea64dbbd14dc1d711c

Former-commit-id: 1467c9e39310fde5cc606f369e6f2a3976eadd22