Fix sharesheet height expansion.
This CL does the following things:
- Add an isIntentResolver check for cases when we want
logic to only apply to ResolverActivity, but not
ChooserActivity.
- Gracefully handle the sticky content preview loading.
Preload it if it will ever be shown, in order to avoid
sudden changes in the share sheet height.
- Don't recalculate the share sheet height when changing
tabs.
- Recalculate the share sheet height if the number
of recycler view children has changed in the cases
when we last calculated it when it wasn't populated yet

Test: manual
Bug: 149817068
Bug: 148156663
Bug: 149823253
Bug: 149732754

Change-Id: If74fc36bd3b8bfa5aa1d91c042fa99e4d19899c9

Former-commit-id: c4dbdbd2e6538ce67eff9caae073eb6221e71c37