Prevent activity config override for system processes.
To improve app compatibility in multi-display and multi-window
environments WM overrides the process config with the config of the
most recently launched activity. However, doing so has caused issues with
system applications (notably System UI) who currently rely on this config.

This change ensures that system processes and system ui do not receive
this compatibility override and revert back to the old functionality. In
the future this will be replaced with dedicated channels to report
display configurations.

Test: WindowProcessControllerTests
Test: ActivityRecordTests

Bug: 148675523
Fixes: 148675523

Change-Id: I811d4f60c1e79b5526a5f9e9973b8d6355ee8c15

Former-commit-id: a098222fb767b4b9919286e2a8d23225d57a6abd