Fix intent resolver height calculation
This CL makes the following changes:
- We update ResolverViewPager to wrap around
the tallest of its children. We can do that now
thanks to recent changes that load the inactive
tab.
- Place empty state screens at the top of the
tab content and update padding and text sizes
- Use layout_alwaysShow=true instead of trying
to calculate the proper height for the different
layout cases (with/without tabs, with/without
default app layout).

Test: manual
Bug: 149823158
Bug: 149822719
Bug: 149823253
Bug: 149817311
Bug: 148156663
Bug: 149824092
Change-Id: I81f9e8600e59b74c64584723054d309becf34f12

Former-commit-id: 3d57e8a8cef274727b701350fa4610c700964528