Track current resized bounds for resized PIP.
Previously, there were only mNormalMovementBounds and
mExpandedMovementBounds since these were the only two sizes PIP window
can be in (on a per-onMovementBoundsChanged call, that is). With the
ability to resize PIP on the horizon, we will need to update the
movement bound as it resizes, and also calculcate this movement with the
new bound on the fly.

Also to make it less confusing, I've renamed:
- normalBounds -> minBounds
- expandedBounds -> maxBounds

Bug: 147361175
Test: Manually. Resize PIP, and then drag around and off-screen, let go
- see PIP window snaps back.
Resize PIP, tap window (see it expands), and wait for timeout. Restores
to previously resized PIP bound.

Change-Id: I2ef286d6fa03914a88ac46e2854548548093e79e

Former-commit-id: 39c9570d459da54aac59a615b9cadd0397383144