Create ACTIVITY_TYPE_DREAM
Previously the dream activity was of a standard activity type. This
allows other activities to be started in front of it, which should not
happen.

This CL adds a new ACTIVITY_TYPE_DREAM and makes dream activities to be
attached on top of all other activities. It also adds activities of type
dream to be one of the always on top activities.

Bug: 150749838

Test: make && flash; start PiP and then start the screensaver. Verify
      that PiP is hidden by the dream.

Change-Id: I852db2455d580d5c99aca3dbbd1ee3bea20aa7ae

Former-commit-id: 45f48d30823ae5c038dfa330eed0f8ee2fa6f71e