Allow location when Car mode is active.
Car mode may be activated without the user plugging the device into the
car. In this scenario, the user is "interacting" with the device via the
car and may need navigation to work. To allow this use case, we make
sure battery saver doesn't disable location when the screen is off if
car mode is enabled. Night mode is also disabled as it could affect the
projected UI.

Bug: 149586834
Test: atest android.os.cts.batterysaving.BatterySaverTest
Test: atest BatterySaverPolicyTest
Test: manual visual confirmation and adb shell dumpsys power scanning
Change-Id: I3390c8a7f94652644f37027f3d22746a6af0efd6

Former-commit-id: 8da3edbb564952f6de1871a2d535f4f98a71a8d0