WindowContainerTransaction: Support PIP Transition
We support the PIP transition by adding setChildWindowingMode
and setWindowingMode primitives. This allows the organizer to
replicate the old behavior during the enter pip animation of
switching the Task to PIP but having the activity continue to
render it's full-screen UI, or the opposite on the other
transition. This is a temporary solution to handle
the boundaries between Fullscreen and PIP states and the trade-off
between TaskOrg and WindowManager. Once TaskOrganizer controls
all windowing modes it won't be needed.

Bug: 149947030
Test: TaskOrganizerTests
Change-Id: I7eeb1b5fb8c5d00a00b6de976f61484f1d067048

Former-commit-id: 2bed621c15871260d7822f18024e6b4ee6b268ff