Deflake NotificationEntryManagerInflationTest
Hopefully deflake NotificationEntryManagerInflationTest by moving back
to the countdown latch approach. This makes us dependent on the listener
interface, but we don't get messed up by random messages on the main
thread. This is likely as good as we get without switching our inflation
pipeline to use an injected executor (or at least off the AsyncTask
API).

Bug: 150618180
Test: atest --iterations 20 NotificationEntryManagerInflationTest
Change-Id: I8c51a2538be38c02ff208fb8aad9a88a10e6c6bf

Former-commit-id: 9c7f4c9810b39d3467856da3e59d290e910eb07e