Comment out the worst flakes (3 tests).
It seems like running shell commands might be the source of the flakes,
but we can't correctly test some functionality without the shell
commands. We need to be able to set properties from a non system package
(e.g. from the shell) to validate that the reset functionality can
distinguish between settings set by the system and those set by anyone
else. While we work on a solution to that, we should disable the most
egregious flakes so that we're not hindering everyone else.

Test: forrest -l remote -b git_master --run_count=100 atest \
      https://googleplex-android-review.git.corp.google.com/c/platform/frameworks/base/+/10383155 \
      SettingsProviderTest:com.android.providers.settings.SettingsProviderTest

Bug: 142206242

Change-Id: I9ccaa3015e08086750cbb1a08262c0742ec1a762

Former-commit-id: 7c4e347847c173d687ad13fe57c0d1a15631098d