Fix screenshot scaling bug
Occasionally, the screenshot preview would not reach the correct
final scale, instead getting "stuck" at some larger size. This is
because there was no final condition for the scale, so the final
state depended on exactly which values the ValueAnimator update
listener was called with. Setting a final state correctly fixes
this problem.

(Also fixes the equivalent bug for the x position.)

Bug: 150134649
Test: manual -- there was not a reliable repro for the bug, but it
hasn't been seen since making the change

Change-Id: Ia383ca9336db382fd54087ce12accafd35b48a22

Former-commit-id: 2750ad0d59551d14eb7667e28b44a713ce3b2c16