Use correct PackageSetting for dexopt calculation
The PackageSetting object is sometimes copied during
scan/install and will not match the reconciledPkg's pkgSetting.

This meant the usesLibrary parsing assigned the values to one
PackageSetting object, but dexopt read from the other, copied
object.

Unfortunately this copy logic leaves the two settings out of
sync and thus the isUpdatedSystemApp flag has to be copied
between them.

All of this would be great to resolve in a future refactor.

Bug: 150142306

Test: adb install -r {out_of_blueline_target}/product/app/CalculatorGooglePrebuilt/CalculatorGooglePrebuilt.apk
    installs correctly, with dexopt run with the correct paths

Change-Id: Ifca3bafcea1b8858baf4a22ddd264d5e07a0f1e5

Former-commit-id: 03f72240fee75d9ef3a8203fabfaf6acdac13244