stubs inheriting from framework-module-stubs-defaults-* now works
This change fixes a problem that droidstub modules inheriting from
framework-module-stubs-defaults-* don't emit any source code. The
problem was happening because

1) --show-annotation was with incorrect format; it had trailing
"," in the parameter list and
2) the class SystemApi was not available to metalava, which made
prevented it from processing any API annotated with @SystemApi.

Fixing the problem by forcibly adding framework annotations to the
sources of the default modules.

Bug: 1241127
Test: `m ike-stubs` and check its output
Change-Id: Icbbbddb4fb68055f27aa314983a9c556e51f037a

Former-commit-id: b4f8a85245d31e57f3f0983158d1e832449cfa4f