Add PlatformCompat permissions to TelephonyProvider
Add the required permissions for querying PlatformCompat to the
whitelist for TelephonyProvider.

This is the minimum scoped permission -- PlatformCompat only defines two
permissions for access (read and write) and we need both.

The alternative for using these permissions is to not call
PlatformCompat and enforce the new behavior on all apps. This was deemed
unacceptable by the app compat team because it was breaking some popular
apps.

Bug: 144631034
Test: atest TelephonyProviderHostTest

Change-Id: I43a0a66ae381caf8d9e21d0f14d471b20a4e44af

Former-commit-id: 106c0defca9abbbd52e3b13c4596015d15082c88