[GWP-ASan] Disable for apps.
Disable GWP-ASan for all applications by default. Long-term
R-release goals are that GWP-ASan is:
 1. Enabled by default for the system.
 2. Enabled by default for system apps.
 3. Disabled by default (with manifest opt-in) for user apps.

At DP2 cut point, we were enabled by default for the system, system
apps, and user apps. We were intending to downgrade user apps to be
disabled-by-default with the manifest changes in ag/10329607, but we
didn't manage to get that through by DP2 cut.

Because we want GWP-ASan to be less aggressive on user apps, it's safest
to just disable GWP-ASan for all apps.

Bug: 135634846
Test: Build and verify no GWP-ASan for apps.
Change-Id: If92cafb49b1484d173a4eadb377d18b29d217cd4

Former-commit-id: 9300488ba6d1d2a72c3badcb0ca2b64c515afe41