Unrestrict updated buggy apps.
If an app is put into the RESTRICTED bucket because the system thinks
it's buggy, we can bring it out of the bucket when the app is updated,
assuming that the bug is fixed in the new version. If there are other
reasons that the app is restricted, then we don't bring it out.

Bug: 149507105
Test: atest FrameworksServicesTests:AppIdleHistoryTests
Test: atest FrameworksServicesTests:AppStandbyControllerTests
Change-Id: Ie4538dde3b367d32f4260e8301afa0556b845b2a

Former-commit-id: 75e1da7bebf476b7a71fd6ec69dc9a6e61712ba1