Add PowerWhitelistManager.getWhitelistedAppIds.
The "XXXExceptIdle" lists and methods are expected to be a superset of
the XXX methods. That's not immediately obvious from the method names.
This effort is to clean up the naming and add documentation so behavior
is better documented. As part of this, we're moving the
getAppIdWhitelisted methods to PowerWhitelistManager to also help
clean up the interface between the apex and the rest of the system.

Bug: 142420609
Bug: 144864180
Bug: 145014493
Test: atest CtsHostsideNetworkTests
Test: atest FrameworksCoreTests:NetworkPolicyManagerTest
Test: atest FrameworksServicesTests:NetworkPolicyManagerServiceTest
Change-Id: I9438a948c916be273f65febd42c3567cc1dc035d

Former-commit-id: 330b74c5b2c8f19dfc8122afa4fe92b182425a2b