Scrub TelephonyRegistry of package names on user builds
For privacy-constrained bug report types (e.g. connectivity), we should
be omitting package names on user builds, but UIDs are acceptable to
display, so we add those in a few more places. For full bug reports on
user builds, the UID can always be mapped back to a package name.

Test: inspect dumpsys telephony.registry on user build
Bug: 150409813
Change-Id: Iad7be2d71a872f48b7384d765b7ffbb9547e2d75

Former-commit-id: bb6c3b562dfde7b8859e5c158e6fc79b519ecfad