Align the end time of wallpaper rotation transform to its target
That fixes the flicker when backing from a different orientation
activity than home. Because the wallpaper window may be updated
later than home's transition. Ideally, the fixed rotation transform
of wallpaper should be synced with its target for consistency.

Also fixes various hidden errors of RecentsAnimationControllerTest.
Mostly about mDisplayContent:
 - It is not default display so it requires additional mock.
 - It contains many testing windows from default setup, that may
   disturb the test flow.
 - Currently it cannot rotate due to wrong condition in
   TestDisplayContent.Builder.
That leads to several false positive results. Now it is changed to
use the existing default display and home stack, which is more
efficient and reasonable.

Bug: 148000485
Test: atest RecentsAnimationControllerTest# \
      testWallpaperHasFixedRotationApplied
Change-Id: I4c054e8a5c9ef654c635055509e2c55f519be932

Former-commit-id: 6393ad12932d4c1a2706ab15fcb7f1e893299e12