Revert "Better preload drawable logging"
Bug: 63136392
Bug: 150395371
Test: make

This reverts commit 078100098a1290740fd0c4d870903869a8a82098 [formerly 1480b67635cdc4c2c5e735741bf30393fd70d738].

Bitmap moving into a new mainline module in S. These module-boundary-
crossing hidden APIs need to be either made official or removed.

We already got the information we need from this logging, so remove it
and the hidden APIs. If/when we need to investigate this in the future,
we should be able to recreate the logging without crossing the boundary.

Change-Id: I5d094a0e427d225719614745867d5b00888d8fec

Former-commit-id: fb654cf97a6f2a224b3f5ce5e411764ba934346a