Revert "Turn on QS media player by default"
This reverts commit 8f69fb92703e23b521c707ae4c4be88a405ec802 [formerly 679609721c4a7422ac3612dde69c9e2639ea6883].

Reason for revert: b/150437753. IM'ed with Beth, going to revert but not merge this CL.  Beth is working on a fix with an eta of a day or 2. Ill CP this into a RP1A.200303.002+ builds.

Change-Id: Id42659d33bbd5ccc188d977f6af320cc8b3b6478
(cherry picked from commit f1563675e81b0b079b90fa3d7c9628c16aab9359)

Former-commit-id: 6f264696e43c55df4a4f221eedad4a4a21b729ad