Prevent intent resolver from changing height when no app targets
If we select the inactive tab, and there are no resolved targets,
the intent resolver changes its height for a few frames.
This is because first we show and reload the list view as we
switch to the other, and then in onPostListReady we show the
empty state screen for no apps available. There are few frames
between these two events, when the list view has a height of 0.
That's when the ResolverViewPager recalculates its height and
thus reduces the intent resolver height. Then as the empty state
screen is shown, the proper intent resolver height is restored.
The fix is to move the showListView call to onPostListReady.

Test: manual
Fixes: 150888674
Change-Id: I685568e183bbd5708242e1228195ec6b09023272

Former-commit-id: 4c0513d8c77bd85dee791cbc1f22ee136f758e33