Copy over PkgState in PackageSetting updateFrom
Previously, the two PackageSetting objects used during
scan/install would reference the same PackageParser.Package
object, so it didn't matter which one was accessed to set
stateful fields. It would implicitly be reflected in the other
PackageSetting through the shared object.

With these fields now moved out of the shared object and
into the PackageSetting itself, these have to be updated
once the 2 settings are reconciled.

Bug: 150812845
Bug: 150596321

Test: manual update a system app, reboot, dumpsys package and
	verify that flags field contains UPDATED_SYSTEM_APP
Test: manual run dexopt for updated system app after install
	and after a reboot; verify paths are correct

Change-Id: Iaf5dfa7fb480eef760f6a44e01f5ee8317055709

Former-commit-id: 10bb870d42d7bacf7670891bc4c0a501e6c88062