Make BootImageProfileTest less flaky
Number one flake is when the system server process is killed before
SIGUSR1. Change this case to be non fatal.

Increase number of retry iterations in case the package manager
is very slow to start and get a JIT sample. Hopefully fixes the
other flake where PackageManagerService.<init> is not seen in the
profile.

Test: atest BootImageProfileTest
Bug: 149802360
Change-Id: I52ac80f15d8efe152dea81b75a82faa4da1b8600

Former-commit-id: d5d442ddd9b59528f84ae84b27b05795c04ef7cb