Take in a Surface then convert to a NativeHandle
We need to pass Surface objects to HIDL, but the Surface type is
currently not supported in HIDL. The closest thing to a Surface in HIDL
is a NativeHandle.

The initial plan was to convert Surface objects to NativeHandle objects
in framework, before passing them to HIDL. Unfortunately, there is no
canonical way to perform the conversion using just the framework APIs.
The only viable approach is to use the NDK conversion functions through
the JNI.

Bug: 150966034
Test: Manual, with a biometrics.face@1.1 HAL
Change-Id: I24f775b576fcab7c0260537d6f9cdc35a4794436

Former-commit-id: e4c6b83572325f2ea441a88c6837664c0e8b014d