Fix up interractions between Adjust for IME and minimize
These aren't compatible at the moment, but at the very least
prevent them from running at the same time so that switching between
secondary adjusted to recents works.

This works by augmenting the divider's ime handling with a paused
state. When paused, it will keep track of ime changes, but it won't
apply them to the divider until it's resumed.

Despite this, the interraction is still quite glitchy, but at least
it ends up in the right end-states. A full fix for this will require
a significant rewrite of the whole DividerView's bounds handling.

Bug: 150781668
Test: Open 2 apps in splits. show IME in secondary, then swipe-up
      to recents. While not perfect, it should work.
Change-Id: I4e7a6b9ed04b38edb83655b2326de4448124b904

Former-commit-id: c0eec05ed584c88c9543c60fdff166f3cbf72682