Synchronize adapter registration with mCallbacks
Move register/unregister callback to the caller thread.
This fixes a problem that if an app registers adapter while the
BluetoothManagerService is turning on Bluetooth, the app would
get a null IBluetooth and would not receive any callback after
Bluetooth is turned on.

Bug: 149890859
Test: Manual
Change-Id: Ia79f6bd1fd6ffab814a15053aeaae031faf875dd
Merged-In: Ia79f6bd1fd6ffab814a15053aeaae031faf875dd

Former-commit-id: 9b0bb63ed7fe0b7d8747f20cd2cc674fd491f3a3