Revert DevicePolicyManager time / zone changes
Revert recent changes that make DevicePolicyManager call TimeDetector /
TimeZoneDetector to change the device time / time zone.

The DPC app runs as the user, so any rules that the TimeDetector /
TimeZoneDetector wants to enforce about what the end user can do will
need to be different for the DPMS path. There will be a dedicated
(probably LocalService) code path for the DPMS to use instead.

Bug: 140712361
Test: treehugger
Change-Id: Ia60702492231cc4c7c5de157c1f266d30996d950

Former-commit-id: 77c9fcdb108218e0267831758963230d361bf1e1