InsetController: Release leashes from RenderThread
We handle changes to the leashes from the UI thread, but use
the same SurfaceControl wrapper object from the RenderThread with
SyncRtSurfaceTransactionApplier. This means that at the time
we release a SurfaceControl from the UI thread we might have already
scheduled a SyncRtSurfaceTransactionApplier to use it, and actually
that could be in the process of running, leading to racy access and
crashes. To fix this we release the SurfaceControl from the RenderThread
so that it happens behind all existing operations.

Bug: 151086678
Test: Existing tests pass.
Change-Id: I2308d1c64f3f368c32587f99ddfb9e05955f821f
(cherry picked from commit 9fa35dea84ae37b53840902992a8884d8d293375 [formerly 521e363a0ff5b90bec8f1fb6d5e10156af713a76])

Former-commit-id: 0eeb8686fa3e3097cc60ab50b61ca58c74e54203