BLASTSyncEngine: Add timeouts
Ensure that if a WindowState fails to call drawFinished
(currently WindowStates are the only level of the hierarchy that
prevent the operation from immediately completing), a timeout
will cause BLAST operations to eventually finish. Without a timeout
it could be pretty hard to handle for the caller, and also leave
us in a bad state if we started new sync operations.

Bug: 148476626
Test: Existing tests pass.
Change-Id: I120fce11532307fadb86ec9086a21e1ca26b8430

Former-commit-id: ae0100bffcb7e9387d88104b9eba02746dc73db8