BLASTSyncEngine: Child Window Support
Add support for syncing WindowState with their children. Recall
that the BLASTSyncEngines work with a 2 way latch:
	1. setReady must be invoked
	2. All the children must invoke transaction ready
So we use a setup where we set-up a sync state from WindowState including
all our children, and then mark it as ready from our own finishDrawing
in order to implement a 2 way latch between us an our children.

Bug: 148871522
Test: TaskOrganizerTests
Change-Id: I14b3160586d06606ac1366b086b8daeba97c8cc2

Former-commit-id: 253765144915806217926293c11f56259fcdeccc