Remove lock inversion in LegacyVpnRunner
This commit prevents the lock inversion introduced in aosp/1218118 by
reverting some of the calls to the cleanup code that ensures all outer
class state is reset. At worst, the correctness of the outer class'
state for the LegacyVpnRunner shutdown is the same as before
aosp/1218118.

This generally seems safe as:
1. This reverts it to a known state (non-ideal, but non-broken)
2. If the interfaces have already been created and registered with the
event observer, the event observer will trigger a interfaceRemoved,
which will both shutdown the LegacyVpnRunner, and cleanup the Vpn
instance state.

Bug: 151068107
Test: FrameworksNetTests passing
Test: Repeated steps in bug, no deadlock
Change-Id: I51a475f575252fb981641fc6240bb8c15576c5f2

Former-commit-id: 144829b1d153eb32f914fa18e1cec8f98471958b