MediaRouter: Notify DiscoveryPreference after binding
When the discovery preference is updated and it's not bound,
the proxy caches the discovery preference and notify the service
when it's bound.

We don't need to pending other commands since we can't do before
the service is bound.(e.g. we don't have a route to select if provider
is not connected)

Bug: 151063538
Bug: 151128724
Test: cts test and manually by sideloading GmsCore
Change-Id: I684806524909c977d6ac246773610899aa3e047a

Former-commit-id: bab0cdad91acf0d7f0f0e271a37ba3a0454f353c