Crash service on SoundTriger HAL fatal error
Throwing an exception is not enough in this case. When the HAL
behaves unexpectedly, the system service and the HAL must be
reset and the client must be notified. Without a full reset in
this catastrophic case, the state of the HAL and the system service
cannot be guaranteed to the client.

Bug: 150569186
Test: Intentionally load an incompatible voice model and confirm the
device can recover.
Change-Id: Ia7c3f25e48f04bf32a96c64ec998fdfa52459685

Former-commit-id: 1e03bc0539308e9b99da1e738d437f196a5e9074