When calculating share sheet height, account for empty state screens.
As we calculate the share sheet height only for the current profile,
there is a rare case when the current profile may have no results.
For that case, we need to adjust the share sheet offset so that
it includes the empty state screen height.

In addition, this CL also does an early return if we are
in the non-main profile (e.g. personal tab in work profile)
to prevent the height from being recalculated upon tab switch.

Fixes: 150939637
Test: manual
Change-Id: Ia465f53d7a2006e1b7c67ad1e95729ed79e2c3a0

Former-commit-id: a30787b8fa76bd81600c8ff3a18fa2090d660e58